[IMP] split files as per each object
