+++ sale_order_lot_selection/models/__init__.py
@@ -1,2 +1,3 @@
-from . import sale
+from . import sale_order
+from . import sale_order_line
 from . import stock
