+++ sale_order_lot_selection/models/sale_order.py
@@ -1,35 +1,5 @@
-from odoo import _, api, fields, models
+from odoo import _, api, models
 from odoo.exceptions import UserError
-
-
-class SaleOrderLine(models.Model):
-    _inherit = "sale.order.line"
-
-    lot_id = fields.Many2one("stock.production.lot", "Lot", copy=False)
-
-    @api.onchange("product_id")
-    def product_id_change(self):
-        super().product_id_change()
-        self.lot_id = False
-
-    @api.onchange("product_id")
-    def _onchange_product_id_set_lot_domain(self):
-        available_lot_ids = []
-        if self.order_id.warehouse_id and self.product_id:
-            location = self.order_id.warehouse_id.lot_stock_id
-            quants = self.env["stock.quant"].read_group(
-                [
-                    ("product_id", "=", self.product_id.id),
-                    ("location_id", "child_of", location.id),
-                    ("quantity", ">", 0),
-                    ("lot_id", "!=", False),
-                ],
-                ["lot_id"],
-                "lot_id",
-            )
-            available_lot_ids = [quant["lot_id"][0] for quant in quants]
-        self.lot_id = False
-        return {"domain": {"lot_id": [("id", "in", available_lot_ids)]}}
 
 
 class SaleOrder(models.Model):
@@ -40,15 +10,14 @@ class SaleOrder(models.Model):
         move = self.env["stock.move"]
         # i create this counter to check lot's univocity on move line
         lot_count = 0
-        for p in line.order_id.picking_ids:
-            for m in p.move_lines:
-                move_line_id = m.move_line_ids.filtered(lambda line: line.lot_id)
-                if move_line_id and line.lot_id == move_line_id[:1].lot_id:
-                    move = m
-                    lot_count += 1
-                    # if counter is 0 or > 1 means that something goes wrong
-                    if lot_count != 1:
-                        raise UserError(_("Can't retrieve lot on stock"))
+        for m in line.order_id.picking_ids.mapped("move_lines"):
+            move_line_id = m.move_line_ids.filtered(lambda line: line.lot_id)
+            if move_line_id and line.lot_id == move_line_id[0].lot_id:
+                move = m
+                lot_count += 1
+                # if counter is 0 or > 1 means that something goes wrong
+                if lot_count != 1:
+                    raise UserError(_("Can't retrieve lot on stock"))
         return move
 
     @api.model
