+++ sale_order_lot_selection/models/stock.py
@@ -16,7 +16,7 @@ class StockMove(models.Model):
     ):
         if self._context.get("sol_lot_id"):
             lot_id = self.sale_line_id.lot_id
-        return super(StockMove, self)._update_reserved_quantity(
+        return super()._update_reserved_quantity(
             need,
             available_quantity,
             location_id,
@@ -27,7 +27,7 @@ class StockMove(models.Model):
         )
 
     def _prepare_move_line_vals(self, quantity=None, reserved_quant=None):
-        vals = super(StockMove, self)._prepare_move_line_vals(
+        vals = super()._prepare_move_line_vals(
             quantity=quantity, reserved_quant=reserved_quant
         )
         if reserved_quant and self.sale_line_id:
